style: rename modules to commands
