+++ pyproject.toml
@@ -24,7 +24,7 @@ source = [
     "wrapper/clients",
     "wrapper/configs",
     "wrapper/engines",
-    "wrapper/modules",
+    "wrapper/commands",
     "wrapper/utils"
 ]
