+++ wrapper/__main__.py
@@ -2,14 +2,13 @@ import os
 import sys
 import json
 import argparse
-from pathlib import Path
 
 import wrapper.tools.cleaning as cm
 import wrapper.tools.messages as msg
 
-from wrapper.modules.bundle_creator import BundleCreator
-from wrapper.modules.kernel_builder import KernelBuilder
-from wrapper.modules.assets_collector import AssetsCollector
+from wrapper.commands.bundle import BundleCreator
+from wrapper.commands.kernel import KernelBuilder
+from wrapper.commands.assets import AssetsCollector
 
 from wrapper.configs import ArgumentConfig
 from wrapper.configs.directory_config import DirectoryConfig as dcfg
@@ -24,7 +23,7 @@ def parse_args() -> argparse.Namespace:
     args = None if sys.argv[1:] else ["-h"]
     # parser and subparsers
     parser_parent = argparse.ArgumentParser(description="A custom wrapper for the zero kernel.")
-    subparsers = parser_parent.add_subparsers(dest="module")
+    subparsers = parser_parent.add_subparsers(dest="command")
     parser_kernel = subparsers.add_parser("kernel", help="build the kernel")
     parser_assets = subparsers.add_parser("assets", help="collect assets")
     parser_bundle = subparsers.add_parser("bundle", help="build the kernel + collect assets")
@@ -242,7 +241,7 @@ def main(args: argparse.Namespace) -> None:
     acfg = ArgumentConfig(**arguments)
     acfg.check_settings()
     # setup output stream
-    if args.module and args.outlog:
+    if args.command and args.outlog:
         msg.note(f"Writing output to {args.outlog}")
         if args.outlog in os.listdir():
             os.remove(args.outlog)
@@ -255,7 +254,7 @@ def main(args: argparse.Namespace) -> None:
         case "podman":
             PodmanEngine(**json.loads(acfg.model_dump_json())).run()
         case "local":
-            match args.module:
+            match args.command:
                 case "kernel":
                     KernelBuilder(
                         codename = acfg.codename,
