+++ wrapper/configs/argument_config.py
@@ -13,7 +13,7 @@ class ArgumentConfig(BaseModel):
     """A variable storage to use across the application.
 
     :param benv: Build environment.
-    :param module: Wrapper module to be launched.
+    :param command: Wrapper command to be launched.
     :param codename: Device codename.
     :param base: Kernel source base.
     :param lkv: Linux kernel version.
@@ -28,7 +28,7 @@ class ArgumentConfig(BaseModel):
     """
 
     benv: str
-    module: str
+    command: str
     codename: str
     base: str
     lkv: Optional[str] = None
@@ -44,7 +44,7 @@ class ArgumentConfig(BaseModel):
     def check_settings(self) -> None:
         """Run settings validations."""
         # allow only asset colletion on a non-Linux machine
-        if self.benv == "local" and self.module in ("kernel", "bundle"):
+        if self.benv == "local" and self.command in ("kernel", "bundle"):
             if not platform.system() == "Linux":
                 msg.error("Can't build kernel on a non-Linux machine.")
             else:
@@ -58,7 +58,7 @@ class ArgumentConfig(BaseModel):
             devices = json.load(f)
         if self.codename not in devices.keys():
             msg.error("Unsupported device codename specified.")
-        if self.module == "bundle":
+        if self.command == "bundle":
             # check Conan-related argument usage
             if self.package_type != "conan" and self.conan_upload:
                 msg.error("Cannot use Conan-related arguments with non-Conan packaging\n")
